smart_research_data_assistant/requirements-full.txt
streamlit>=1.39.0
pandas>=2.1.0,<2.2.0
numpy>=1.26.0
plotly>=5.24.0
matplotlib>=3.7.0,<3.8.0
scikit-learn>=1.4.0,<1.6.0
pycaret>=3.3.0,<3.4.0
tensorflow>=2.17.0
torch>=2.4.0
torchvision>=0.19.0
opencv-python-headless>=4.10.0
pillow>=10.4.0
nltk>=3.9.0
spacy>=3.7.0
openai>=1.51.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
faiss-cpu>=1.8.0
chromadb>=0.5.0
pypdf>=5.0.0
python-dotenv>=1.0.1
openpyxl>=3.1.0
sentence-transformers>=3.1.0
ultralytics>=8.3.0
langchain-huggingface>=0.1.0

